requirements.txt
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic-settings==2.9.1
python-dotenv==1.1.0
pytest==8.2.1